+++ oceanics-io-wasm-ui/src/grid/mini_map.rs
@@ -1,0 +1,3 @@
+pub mod mini_map {
+    
+}
